Switch to using subprocess.run()
The preivos combination was prone to deadlocking when there wa a lot of
output, which the new code that runs migration was much more likely to
have.

Since python3 provides a much more convenient function in
subprocess.run(), just switch to using that one.

While at it, increase timeout to 30 seocnds as well, since migrations
can take more time.